+++ CHANGELOG.md
@@ -14,6 +14,7 @@ app.
 - (Firefox) Fixed occasional flicker on loading Web pages over slow connections
   when the extension is enabled.
 - Fixed lookup of Japanese inside `display: contents` containers.
+- Made the copy overlay not show if the user has selected text in the popup.
 
 ## [1.12.5] - 2022-09-17